pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarmx"
version = "0.1.0"
description = "A model-agnostic, async, event-driven multi-agent orchestration framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "SwarmX Contributors" },
]
keywords = ["agents", "orchestration", "multi-agent", "async", "event-driven", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
swarmx = "swarmx.cli.main:cli"

[project.urls]
Homepage = "https://github.com/swarmx/swarmx"
Repository = "https://github.com/swarmx/swarmx"
Issues = "https://github.com/swarmx/swarmx/issues"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
